refactor: 페이지 내 공통된 스타일 분리
