+++ frontend/src/pages/Admin/CreateCoupon/CreateCoupon.style.tsx
@@ -1,32 +1,7 @@
 import { styled } from 'styled-components';
+import { PageContainer } from '../../../style/layout/common';
 
-interface StyledSpacingProps {
-  $size: number;
-}
-
-export const CreateCouponContainer = styled.div`
-  margin: 0 auto;
-`;
-
-export const Spacing = styled.div<StyledSpacingProps>`
-  width: 100%;
-  height: ${({ $size }) => `${$size}px`};
-`;
-
-export const Title = styled.h2`
-  font-size: 36px;
-  font-weight: 700;
-`;
-
-export const SubTitle = styled.p`
-  font-size: 16px;
-  font-weight: 700;
-`;
-
-export const Text = styled.span`
-  font-size: 12px;
-  color: ${({ theme }) => theme.colors.black};
-`;
+export const CreateCouponContainer = styled(PageContainer)``;
 
 export const NextButtonWrapper = styled.div`
   display: flex;
