+++ frontend/src/pages/Admin/CreateCoupon/index.tsx
@@ -1,15 +1,10 @@
 import { useRef, useState } from 'react';
 import Button from '../../../components/Button';
 import { Input } from '../../../components/Input';
-import {
-  CreateCouponContainer,
-  NextButtonWrapper,
-  Spacing,
-  SubTitle,
-  Text,
-  Title,
-} from './CreateCoupon.style';
+import { CreateCouponContainer, NextButtonWrapper } from './CreateCoupon.style';
 import RadioInputs from './RadioInputs';
+import { useNavigate } from 'react-router-dom';
+import { Spacing, Title, SectionTitle, Text } from '../../../style/layout/common';
 
 const ParagraphSpacing = () => <Spacing $size={8} />;
 
@@ -21,16 +16,17 @@ const CreateCoupon = () => {
   const rewardInputRef = useRef<HTMLInputElement>(null);
   const expireSelectedRef = useRef<HTMLInputElement>(null);
   const stampCountRef = useRef<HTMLInputElement>(null);
+  const navigate = useNavigate();
 
   return (
     <CreateCouponContainer>
       <Title>쿠폰 제작 및 변경</Title>
       <SectionSpacing />
-      <SubTitle>어떻게 제작하시겠어요?</SubTitle>
+      <SectionTitle>어떻게 제작하시겠어요?</SectionTitle>
       <ParagraphSpacing />
       <RadioInputs setValue={setCreatedType} />
       <SectionSpacing />
-      <SubTitle>목표 스탬프 갯수</SubTitle>
+      <SectionTitle>목표 스탬프 갯수</SectionTitle>
       <ParagraphSpacing />
       <Text>리워드를 지급할 스탬프 갯수를 작성해주세요.</Text>
       {/* TODO: placeholder가 적용되지 않는 오류가 있음. */}
@@ -44,7 +40,7 @@ const CreateCoupon = () => {
         ref={rewardInputRef}
       />
       <SectionSpacing />
-      <SubTitle>쿠폰 유효기간</SubTitle>
+      <SectionTitle>쿠폰 유효기간</SectionTitle>
       <ParagraphSpacing />
       <Text>
         고객이 가지게 될 쿠폰의 유효기간입니다. 유효기간이 지나면 해당 쿠폰의 스탬프가 모두
@@ -52,7 +48,11 @@ const CreateCoupon = () => {
       </Text>
       <SectionSpacing />
       <NextButtonWrapper>
-        <Button variant="secondary" size="medium">
+        <Button
+          variant="secondary"
+          size="medium"
+          onClick={() => navigate('/admin/create-coupon/2')}
+        >
           다음으로
         </Button>
       </NextButtonWrapper>
